% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.R
\name{af}
\alias{af}
\title{Present Value of An Annuity}
\usage{
af(l = 0, n, i)
}
\arguments{
\item{l}{0 for annuity due or 1 for annuity immediate.}

\item{n}{A numeric value. The number of payments.}

\item{i}{A numeric value. The interest rate.}
}
\description{
Calculates the present value of an annuity.
}
\examples{
af(0,10,0.03)
af(1,15,0.05)

}
\keyword{Annuity}
\keyword{Present}
\keyword{Value}
