% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgentinaINDEC9092comb.R
\docType{data}
\name{ArgentinaINDEC9092comb}
\alias{ArgentinaINDEC9092comb}
\title{ArgentinaINDEC9092 Males and Females Combined}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(ArgentinaINDEC9092comb)
}
\description{
Mortality table (ultimate): Argentina Instituto Nacional de Estadistica y Censos (INDEC). Nation:	Argentina. Year: 1990-1992. Sex: Males and Females Combined.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20003}
}
\keyword{datasets}
\keyword{table}
