% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgentinaINDEC9092F.R
\docType{data}
\name{ArgentinaINDEC9092F}
\alias{ArgentinaINDEC9092F}
\title{ArgentinaINDEC9092 Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(ArgentinaINDEC9092F)
}
\description{
Mortality table (ultimate): Argentina Instituto Nacional de Estadistica y Censos (INDEC). Nation:	Argentina. Year: 1990-1992. Sex: Female.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20002}
}
\keyword{datasets}
\keyword{table}
