% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgentinaINDEC9092M.R
\docType{data}
\name{ArgentinaINDEC9092M}
\alias{ArgentinaINDEC9092M}
\title{ArgentinaINDEC9092 Male}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(ArgentinaINDEC9092M)
}
\description{
Mortality table (ultimate): Argentina Instituto Nacional de Estadistica y Censos (INDEC). Nation:	Argentina. Year: 1990-1992. Sex: Male.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20001}
}
\keyword{datasets}
\keyword{table}
