% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contmeses.R
\name{contmeses}
\alias{contmeses}
\title{Contmeses}
\usage{
contmeses(x, var)
}
\arguments{
\item{x}{A numeric type value. The moment.}

\item{var}{A numeric type value. Number of payments per year.}
}
\value{
Returns 1 if in that moment corresponds a payment or 0 if not.
}
\description{
Helpful function for identifying if a payment is required on that period.
}
\examples{
DetLifeInsurance:::contmeses(13,2)

}
\keyword{Contmeses}
