% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO2001FALBnonsmoker.R
\docType{data}
\name{CSO2001FALBnonsmoker}
\alias{CSO2001FALBnonsmoker}
\title{CSO2001 Female Age Last Birthday Non-smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO2001FALBnonsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation: United States of America. Year: 2001. Sex: Female. Basis: Age Last Birthday. Smoker: No.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1517}
}
\keyword{datasets}
\keyword{table}
