% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO2001FALBsmoker.R
\docType{data}
\name{CSO2001FALBsmoker}
\alias{CSO2001FALBsmoker}
\title{CSO2001 Female Age Last Birthday Smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO2001FALBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation: United States of America. Year: 2001. Sex: Female. Basis: Age Last Birthday. Smoker: yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1519}
}
\keyword{datasets}
\keyword{table}
