% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO2001MALBsmoker.R
\docType{data}
\name{CSO2001MALBsmoker}
\alias{CSO2001MALBsmoker}
\title{CSO2001 Male Age Last Birthday Smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO2001MALBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation: United States of America. Year: 2001. Sex: Male. Basis: Age Last Birthday. Smoker: yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1518}
}
\keyword{datasets}
\keyword{table}
