% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO2001MANBnonsmoker.R
\docType{data}
\name{CSO2001MANBnonsmoker}
\alias{CSO2001MANBnonsmoker}
\title{CSO2001 Male Age Nearest Birthday Non-smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO2001MANBnonsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation: United States of America. Year: 2001. Sex: Male. Basis: Age Nearest Birthday. Smoker: No.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1137}
}
\keyword{datasets}
\keyword{table}
