% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO2001MANBsmoker.R
\docType{data}
\name{CSO2001MANBsmoker}
\alias{CSO2001MANBsmoker}
\title{CSO2001 Male Age Nearest Birthday Smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO2001MANBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation: United States of America. Year: 2001. Sex: Male. Basis: Age Nearest Birthday. Smoker: Yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1138}
}
\keyword{datasets}
\keyword{table}
