% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO58FALB.R
\docType{data}
\name{CSO58FALB}
\alias{CSO58FALB}
\title{CSO58 Female Age Last Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO58FALB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Year: 1958. Nation: United States of America. Sex: Female. Basis: Age Last Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=8}
}
\keyword{datasets}
\keyword{table}
