% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO58FANB.R
\docType{data}
\name{CSO58FANB}
\alias{CSO58FANB}
\title{CSO58 Female Age Nearest Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO58FANB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1958. Sex: Female. Basis: Age Nearest Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=6}
}
\keyword{datasets}
\keyword{table}
