% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO58MALB.R
\docType{data}
\name{CSO58MALB}
\alias{CSO58MALB}
\title{CSO58 Male Age Last Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO58MALB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1958. Sex: Male. Basis: Age Last Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=7}
}
\keyword{datasets}
\keyword{table}
