% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO58MANB.R
\docType{data}
\name{CSO58MANB}
\alias{CSO58MANB}
\title{CSO58 Male Age Nearest Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO58MANB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1958. Sex: Male. Basis: Age Nearest Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=5}
}
\keyword{datasets}
\keyword{table}
