% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80FALB.R
\docType{data}
\name{CSO80FALB}
\alias{CSO80FALB}
\title{CSO80 Female Age Last Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80FALB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation: United States of America. Year: 1980. Sex: Female Age method: Age Last Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=35}
}
\keyword{datasets}
\keyword{table}
