% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80FALBnonsmoker.R
\docType{data}
\name{CSO80FALBnonsmoker}
\alias{CSO80FALBnonsmoker}
\title{CSO80 Female Age Last Birthday non-smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80FALBnonsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Female. Basis: Age Last Birthday. Smoker: No.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=37}
}
\keyword{datasets}
\keyword{table}
