% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80FALBsmoker.R
\docType{data}
\name{CSO80FALBsmoker}
\alias{CSO80FALBsmoker}
\title{CSO80 Female Age Last Birthday smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80FALBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Female. Basis: Age Last Birthday. Smoker: Yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=39}
}
\keyword{datasets}
\keyword{table}
