% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80FANB.R
\docType{data}
\name{CSO80FANB}
\alias{CSO80FANB}
\title{CSO80 Female Age Nearest Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80FANB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Female. Basis: Age Nearest Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=36}
}
\keyword{datasets}
\keyword{table}
