% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80FANBnonsmoker.R
\docType{data}
\name{CSO80FANBnonsmoker}
\alias{CSO80FANBnonsmoker}
\title{CSO80 Female Age Nearest Birthday Non-smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80FANBnonsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Female. Basis: Age Nearest Birthday. Smoker: No.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=38}
}
\keyword{datasets}
\keyword{table}
