% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80FANBsmoker.R
\docType{data}
\name{CSO80FANBsmoker}
\alias{CSO80FANBsmoker}
\title{CSO80 Female Age Nearest Birthday Smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80FANBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Female. Basis: Age Nearest Birthday. Smoker: Yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=40}
}
\keyword{datasets}
\keyword{table}
