% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80MALB.R
\docType{data}
\name{CSO80MALB}
\alias{CSO80MALB}
\title{CSO80 Male Age Last Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80MALB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Male. Basis: Age Last Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=41}
}
\keyword{datasets}
\keyword{table}
