% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80MALBnonsmoker.R
\docType{data}
\name{CSO80MALBnonsmoker}
\alias{CSO80MALBnonsmoker}
\title{CSO80 Male Age Last Birthday Non-smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80MALBnonsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Male. Basis: Age Last Birthday. Smoker: No.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=43}
}
\keyword{datasets}
\keyword{table}
