% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80MALBsmoker.R
\docType{data}
\name{CSO80MALBsmoker}
\alias{CSO80MALBsmoker}
\title{CSO80 Male Age Last Birthday Smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80MALBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Male. Basis: Age Last Birthday. Smoker: Yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=45}
}
\keyword{datasets}
\keyword{table}
