% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80MANB.R
\docType{data}
\name{CSO80MANB}
\alias{CSO80MANB}
\title{CSO80 Male Age Nearest Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80MANB)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Male. Age method: Age Nearest Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=42}
}
\keyword{datasets}
\keyword{table}
