% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80MANBnonsmoker.R
\docType{data}
\name{CSO80MANBnonsmoker}
\alias{CSO80MANBnonsmoker}
\title{CSO80 Male Age Nearest Birthday Non-smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80MANBnonsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Male. Basis: Age Nearest Birthday. Smoker: No.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=44}
}
\keyword{datasets}
\keyword{table}
