% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSO80MANBsmoker.R
\docType{data}
\name{CSO80MANBsmoker}
\alias{CSO80MANBsmoker}
\title{CSO80 Male Age Nearest Birthday Smoker}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(CSO80MANBsmoker)
}
\description{
Mortality table (ultimate): Commissioner's Standard Ordinary. Nation:	United States of America. Year: 1980. Sex: Male. Basis: Age Nearest Birthday. Smoker: Yes.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=46}
}
\keyword{datasets}
\keyword{table}
