% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM71F.R
\docType{data}
\name{GAM71F}
\alias{GAM71F}
\title{GAM71 Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(GAM71F)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Nation: United States of America. Year: 1971. Sex: Female.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=817}, \url{http://servicios.infoleg.gob.ar/infolegInternet/anexos/80000-84999/81029/norma.htm}
}
\keyword{datasets}
\keyword{table}
