% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM83F.R
\docType{data}
\name{GAM83F}
\alias{GAM83F}
\title{GAM83 Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(GAM83F)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Nation: United States of America. Year: 1983. Sex: Female.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=825}
}
\keyword{datasets}
\keyword{table}
