% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM83M.R
\docType{data}
\name{GAM83M}
\alias{GAM83M}
\title{GAM83 Male}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(GAM83M)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Nation: United States of America. Year: 1983. Sex: Male.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=826}
}
\keyword{datasets}
\keyword{table}
