% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM94F.R
\docType{data}
\name{GAM94F}
\alias{GAM94F}
\title{GAM94 Female}
\format{
a dataframe containing a column for age (x) and a column for death probability (q)
}
\usage{
data(GAM94F)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Year: 1994. Sex: Female.
}
\references{
\url{https://mort.soa.org/}
}
\keyword{datasets}
\keyword{table}
