% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM94FANB.R
\docType{data}
\name{GAM94FANB}
\alias{GAM94FANB}
\title{GAM94 Female Age Nearest Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(GAM94FANB)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Nation: United States of America. Year: 1994. Sex: Female. Basis: Age Nearest Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=834}
}
\keyword{datasets}
\keyword{table}
