% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM94M.R
\docType{data}
\name{GAM94M}
\alias{GAM94M}
\title{GAM94 Male}
\format{
a dataframe containing a column for age (x) and a column for death probability (q)
}
\usage{
data(GAM94M)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Year: 1994. Sex: Male.
}
\references{
\url{https://mort.soa.org/}
}
\keyword{datasets}
\keyword{table}
