% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM94MANB.R
\docType{data}
\name{GAM94MANB}
\alias{GAM94MANB}
\title{GAM94 Male Age Nearest Birthday}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(GAM94MANB)
}
\description{
Mortality table (ultimate): Group Annuity Mortality. Nation: United States of America. Year: 1994. Sex: Male. Basis: Age Nearest Birthday.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=835}
}
\keyword{datasets}
\keyword{table}
