% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.R
\name{is_integer}
\alias{is_integer}
\title{Integer}
\usage{
is_integer(x)
}
\arguments{
\item{x}{A numeric type value or a vector.}
}
\description{
Returns 1 if the number or elements of a vector is integer or 0 if it's not.
}
\examples{
DetLifeInsurance:::is_integer(3)
DetLifeInsurance:::is_integer(1:7)

}
\keyword{Integer}
