% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointSurvival.R
\name{JointSurvival}
\alias{JointSurvival}
\title{Joint Survival Probability}
\usage{
JointSurvival(x, n, data, prop = 1)
}
\arguments{
\item{x}{A vector representing the age of each individual.}

\item{n}{An integer. The term.}

\item{data}{A data.frame of the mortality table, with the first column being the age and the second one, the probability of death.}

\item{prop}{A numeric value. The proportion of the mortality table used, between 0 and 1.}
}
\description{
Calculates the probability of survival given a mortality table for a group.
}
\examples{
ages<-c(34,45,52,65)
JointSurvival(ages,10,CSO80FALB)

}
\keyword{Joint}
\keyword{Probability}
\keyword{Survival}
