% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAyP0206activeF.R
\docType{data}
\name{MAyP0206activeF}
\alias{MAyP0206activeF}
\title{MAyP0206 Active Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(MAyP0206activeF)
}
\description{
Mortality table (ultimate): Mortalidad Activos y Pasivos. Nation:	Argentina. Year: 2002-2006. Sex: Female. Status: Active.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20005}
}
\keyword{datasets}
\keyword{table}
