% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAyP0206activeM.R
\docType{data}
\name{MAyP0206activeM}
\alias{MAyP0206activeM}
\title{MAyP0206 Active Male}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(MAyP0206activeM)
}
\description{
Mortality table (ultimate): Mortalidad Activos y Pasivos. Nation:	Argentina. Year: 2002-2006. Sex: Male. Status: Active.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20004}
}
\keyword{datasets}
\keyword{table}
