% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAyP0206CAF.R
\docType{data}
\name{MAyP0206CAF}
\alias{MAyP0206CAF}
\title{MAyP0206 Combined Active and Retired Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(MAyP0206CAF)
}
\description{
Mortality table (ultimate): Mortalidad Activos y Pasivos. Nation:	Argentina. Year: 2002-2006. Sex: Female. Status: Combined Active and Retired.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20009}
}
\keyword{datasets}
\keyword{table}
