% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAyP0206CAM.R
\docType{data}
\name{MAyP0206CAM}
\alias{MAyP0206CAM}
\title{MAyP0206 Combined Active and Retired Male}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(MAyP0206CAM)
}
\description{
Mortality table (ultimate): Mortalidad Activos y Pasivos. Nation:	Argentina. Year: 2002-2006. Sex: Male. Status: Combined Active and Retired.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20008}
}
\keyword{datasets}
\keyword{table}
