% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAyP0206retiredF.R
\docType{data}
\name{MAyP0206retiredF}
\alias{MAyP0206retiredF}
\title{MAyP0206 Retired Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(MAyP0206retiredF)
}
\description{
Mortality table (ultimate): Mortalidad Activos y Pasivos. Nation:	Argentina. Year: 2002-2006. Sex: Female. Status: Retired.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20007}
}
\keyword{datasets}
\keyword{table}
