% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAyP0206retiredM.R
\docType{data}
\name{MAyP0206retiredM}
\alias{MAyP0206retiredM}
\title{MAyP0206 Retired Male}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(MAyP0206retiredM)
}
\description{
Mortality table (ultimate): Mortalidad Activos y Pasivos. Nation:	Argentina. Year: 2002-2006. Sex: Male. Status: Retired.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=20006}
}
\keyword{datasets}
\keyword{table}
