% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mi06F.R
\docType{data}
\name{Mi06F}
\alias{Mi06F}
\title{Mi06 Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(Mi06F)
}
\description{
Mortality table (ultimate): Mortalidad Invalidez. Nation:	Chile. Year: 2006. Sex: Female.
}
\note{
for more information on how to adjust the values of the table using an 'improvement rate' visit: \url{https://www.spensiones.cl/portal/compendio/596/w3-propertyvalue-3537.html}
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=2713},\url{https://www.spensiones.cl/portal/compendio/596/w3-propertyvalue-3542.html}
}
\keyword{datasets}
\keyword{table}
