% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mi85F.R
\docType{data}
\name{Mi85F}
\alias{Mi85F}
\title{Mi85 Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(Mi85F)
}
\description{
Mortality table (ultimate): Mortalidad Invalidez. Nation: Chile. Year: 1985. Sex: Female.
}
\references{
\url{http://servicios.infoleg.gob.ar/infolegInternet/anexos/80000-84999/81029/norma.htm}
}
\keyword{datasets}
\keyword{table}
