% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV04F.R
\docType{data}
\name{RV04F}
\alias{RV04F}
\title{RV04 Female}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(RV04F)
}
\description{
Mortality table (ultimate): Renta Vitalicia. Nation: Chile. Year: 2004. Sex: Female.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1500}
}
\keyword{datasets}
\keyword{table}
