% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV04M.R
\docType{data}
\name{RV04M}
\alias{RV04M}
\title{RV04 Male}
\format{
A data frame containing a column for age (x) and a column for death probability (q).
}
\usage{
data(RV04M)
}
\description{
Mortality table (ultimate): Renta Vitalicia. Nation:	Chile. Year: 2004. Sex: Male.
}
\references{
\url{https://mort.soa.org/ViewTable.aspx?&TableIdentity=1499}
}
\keyword{datasets}
\keyword{table}
