% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf}
\alias{sf}
\title{Future Value of an Annuity}
\usage{
sf(l = 0, n, i)
}
\arguments{
\item{l}{0 for annuity due or 1 for annuity immediate.}

\item{n}{A numeric value. The number of payments.}

\item{i}{A numeric value. The interest rate.}
}
\description{
Calculates the future value  of an annuity.
}
\examples{
sf(0,12,0.05)
sf(1,23,0.04)

}
\keyword{Annuity}
\keyword{Future}
\keyword{Value}
