% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table_Dormoy.R
\name{Table_Dormoy}
\alias{Table_Dormoy}
\title{Dormoy's Law of Mortality Table Creator}
\usage{
Table_Dormoy(x0, omega, a)
}
\arguments{
\item{x0}{A numeric type value. The initial age of the table.}

\item{omega}{A numeric type value. The final age of the table.}

\item{a}{A numeric type value. A parameter of the law.}
}
\value{
Returns a data.frame object containing age and death probabilities.
}
\description{
Creates a mortality table under Dormoy's law.
}
\examples{
Table_Dormoy(0,100,0.98)

}
\references{
Chapter 3 (p 77-78) of  Actuarial Mathematics (1997) by Bowers, Gerber, Hickman, Jones & Nesbitt.
}
\keyword{Dormoy's}
\keyword{law}
\keyword{mortality}
\keyword{table.}
