% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_av.R
\name{V_av}
\alias{V_av}
\title{Reserve Valuation for Varying Life Annuities: Arithmetic Progression}
\usage{
V_av(
  px,
  x,
  h,
  n,
  k = 1,
  r,
  cantprem = 1,
  premperyear = 1,
  i = 0.04,
  data,
  prop = 1,
  assumption = "none",
  variation = "none",
  cap,
  t
)
}
\arguments{
\item{px}{A numeric value. The value of the premium paid in each period.}

\item{x}{An integer. The age of the insuree.}

\item{h}{An integer. The deferral period.}

\item{n}{An integer. Number of years of coverage.}

\item{k}{An integer. Number of payments per year.}

\item{r}{The variation rate. A numeric type value.}

\item{cantprem}{An integer. The total number of premiums.}

\item{premperyear}{An integer. The number of premiums to be paid per year.}

\item{i}{The interest rate. A numeric type value.}

\item{data}{A data.frame containing the mortality table, with the first column being the age and the second one, the probability of death.}

\item{prop}{A numeric value. It represents the proportion of the mortality table used (between 0 and 1).}

\item{assumption}{A character string. The assumption used for fractional ages ("UDD" for uniform distribution of deaths, "constant" for constant force of mortality and "none" if there is no fractional coverage).}

\item{variation}{A character string. "inter" if the variation it's interannual or "intra" if it's intra-annual.}

\item{cap}{A numeric type value. The annualized value of the first payment.}

\item{t}{An integer. The moment of valuation (in months if it is a fractional coverage or in years if it is not).}
}
\value{
A data frame with Premium, Risk, 1/E and reserve values up to the moment t.
}
\description{
Calculates the reserve for the Varying Life Annuity up to the moment t.
}
\examples{
V_av(9435943.49607651,20,2,2,1,0.05,2,1,0.04,CSO80MANB,1,"none","none",10000000,4)
V_av(9516712.17583443/12,20,2,2,2,0.05,24,12,0.04,CSO80MANB,1,"constant","inter",10000000,48)
V_av(9517.04683383614/12,20,2,2,2,0.05,24,12,0.04,CSO80MANB,1,"UDD","inter",10000,48)
V_av(997.404109454868/12,20,2,2,2,0.05,24,12,0.04,CSO80MANB,1,"constant","intra",1000,48)
V_av(997436.738989113/12,20,2,2,2,0.05,24,12,0.04,CSO80MANB,1,"UDD","intra",1000000,48)
V_av(28.4421691213902,40,3,7,2,0.7,1,1,0.04,CSO80MANB,1,"UDD","intra",1,120)

}
\references{
Chapter 5 of  Life Contingencies (1952) by Jordan, Chapter 11 of  Actuarial Mathematics for Life Contingent Risks (2009) by Dickson, Hardy and Waters.
}
\keyword{Annuities}
\keyword{Arithmetic}
\keyword{Life}
\keyword{Progression}
\keyword{Reserve}
\keyword{Varying}
